feat: Added support for PayPal payments
Front end:

 - Integrated PayPal's JavaScript SDK into the frontend, which involved
making changes to several files such as the checkout component,
ecommerce, and shopping cart service. Additionally, I added new files
including environment, PayPal orders, and PayPal service files.

 - In the next commit, I plan to remove the checkout HTML since PayPal
takes care of all the form processing.

Backend:

 - Added support for using PayPal's REST APIs, which required creating
several DTO classes to represent JSON objects coming from and going to
PayPal's request endpoints.

 - Created a PayPal service and controller to handle these requests and
responses.

 - Added PayPal variables to the application.properties file for use in
the service layer, but currently, I'm not able to use them. Instead, I
manually added the variables to the code (not best practice), but it
works for a personal project.

 This commit represents a significant milestone in implementing PayPal
payments for my project. The added support for PayPal payments is a
major improvement. However there is still much to add

+++ springboot/src/main/java/com/example/ecommerceapp/entity/CustomerCart.java
@@ -44,7 +44,7 @@ public class CustomerCart {
     		inverseJoinColumns = @JoinColumn(name="product_id")
     		)
     private List<Product> products;
-    private Integer totalCost;
+    private Float totalCost;
     private LocalDateTime dateCreated;
     private LocalDateTime dateUpdated;
